🔵 move function to request.go file
